Parse SAML group mapping stringified array
See #1765

If you use Cognito to translate from multiple SAML roles to an
user attribute, tokens will include a stringified array, not an
actual array. There is a feature request with AWS to fix this,
but no current ETA.

Solve the issue by identifying string claims of the form "[A, B, ... ]"
and converting that to an array. Right now, the separator for
multiple entries is hardcoded to ", ", but hopefully Cognito in
this particular configuration is the only provider to exhibit this
broken/unfortunate behaviour and there won't be need to support
other variants.